app/requirements.txt
streamlit
google-genai
cryptography
pydantic
requests
jinja2
sentry-sdk
markdown2
psycopg2-binary